requirements.txt
paho-mqtt==2.1.0
Flask==3.1.1
python-dotenv==1.1.1
redis==5.0.4